+++ PropertyManagement/app/src/main/java/com/example/property_management/data/Property.java
@@ -7,8 +7,8 @@ public class Property {
     private String propertyId;
     private String href;
     private String address;
-    private float lat;
-    private float lng;
+    private double lat;
+    private double lng;
     private int numBedrooms;
     private int numBathrooms;
     private int numParking;
@@ -17,7 +17,7 @@ public class Property {
     private ArrayList<String> images;
 
     public Property() {};
-    public Property(String propertyId, String href, String address, float lat, float lng, int numBedrooms, int numBathrooms, int numParking, HashMap<String, RoomData> propertyData, ArrayList<String> images, int price) {
+    public Property(String propertyId, String href, String address, double lat, double lng, int numBedrooms, int numBathrooms, int numParking, HashMap<String, RoomData> propertyData, ArrayList<String> images, int price) {
         this.propertyId = propertyId;
         this.href = href;
         this.address = address;
@@ -43,13 +43,13 @@ public class Property {
 
     public void setAddress(String address) { this.address = address; }
 
-    public float getLat() { return lat; }
+    public double getLat() { return lat; }
 
-    public void setLat(float lat) { this.lat = lat; }
+    public void setLat(double lat) { this.lat = lat; }
 
-    public float getLng() { return lng; }
+    public double getLng() { return lng; }
 
-    public void setLng(float lng) { this.lng = lng; }
+    public void setLng(double lng) { this.lng = lng; }
 
     public int getNumBedrooms() { return numBedrooms; }
