+++ PropertyManagement/app/src/main/java/com/example/property_management/ui/activities/TestActivity.java
@@ -86,37 +86,37 @@ public class TestActivity extends AppCompatActivity {
         properties.add(new Property("0000000000000",
                 "https://www.example1.com",
                 "11 Charming Street Hampton East VIC 3188",
-                (float) 144.022, (float) 132.102,
+                (double) 144.022, (double) 132.102,
                 4, 2, 1,
                 null, images, 0));
         properties.add(new Property("1111111111111",
                 "https://www.example2.com",
                 "12 Charming Street Hampton East VIC 3188",
-                (float) 143.022, (float) 132.102,
+                (double) 143.022, (double) 132.102,
                 2, 1, 0,
                 null, images, 0));
         properties.add(new Property("2222222222222",
                 "https://www.example3.com",
                 "13 Charming Street Hampton East VIC 3188",
-                (float) 145.022, (float) 132.102,
+                (double) 145.022, (double) 132.102,
                 3, 2, 2,
                 null, null, 0));
         properties.add(new Property("0000000000000",
                 "https://www.example1.com",
                 "11 Charming Street Hampton East VIC 3188",
-                (float) 144.022, (float) 132.102,
+                (double) 144.022, (double) 132.102,
                 4, 2, 1,
                 null, images, 0));
         properties.add(new Property("1111111111111",
                 "https://www.example2.com",
                 "12 Charming Street Hampton East VIC 3188",
-                (float) 143.022, (float) 132.102,
+                (double) 143.022, (double) 132.102,
                 2, 1, 0,
                 null, images, 0));
         properties.add(new Property("2222222222222",
                 "https://www.example3.com",
                 "13 Charming Street Hampton East VIC 3188",
-                (float) 145.022, (float) 132.102,
+                (double) 145.022, (double) 132.102,
                 3, 2, 2,
                 null, null, 0));
